SIT-1185 feedback attachment support

+++ app/controllers/AttachmentController.java
@@ -5,6 +5,7 @@ import be.objectify.deadbolt.java.actions.Group;
 import be.objectify.deadbolt.java.actions.Restrict;
 import com.avaje.ebean.Ebean;
 import models.Attachment;
+import models.Comment;
 import models.Exam;
 import models.questions.Answer;
 import models.questions.Question;
@@ -66,14 +67,23 @@ public class AttachmentController extends BaseController {
         } catch (IOException e) {
             return internalServerError("sitnet_error_creating_attachment");
         }
+        // Remove existing one if found
+        Answer answer = question.getAnswer();
+        if (answer.getAttachment() != null) {
+            Attachment aa = answer.getAttachment();
+            answer.setAttachment(null);
+            answer.save();
+            AppUtil.removeAttachmentFile(aa.getFilePath());
+            aa.delete();
+        }
 
         Attachment attachment = new Attachment();
         attachment.setFileName(fileName);
         attachment.setFilePath(newFilePath);
         attachment.setMimeType(contentType);
         attachment.save();
-        question.getAnswer().setAttachment(attachment);
-        question.save();
+        answer.setAttachment(attachment);
+        answer.save();
         return ok(attachment);
     }
 
@@ -101,6 +111,13 @@ public class AttachmentController extends BaseController {
         } catch (IOException e) {
             return internalServerError("sitnet_error_creating_attachment");
         }
+        // Remove existing one if found
+        if (question.getAttachment() != null) {
+            Attachment aa = question.getAttachment();
+            question.setAttachment(null);
+            question.save();
+            aa.delete();
+        }
 
         Attachment attachment = new Attachment();
         attachment.setFileName(filePart.getFilename());
@@ -118,30 +135,28 @@ public class AttachmentController extends BaseController {
     public Result deleteQuestionAttachment(Long id) {
 
         Question question = Ebean.find(Question.class, id);
-        Attachment aa = Ebean.find(Attachment.class, question.getAttachment().getId());
-
-        question.setAttachment(null);
-        question.save();
-
-        aa.delete();
-        // DO NOT DELETE THE ACTUAL FILE, IT MAY BE REFERENCED FROM CHILD QUESTIONS!
-        //AppUtil.removeAttachmentFile(aa.getFilePath());
-
+        if (question != null && question.getAttachment() != null) {
+            Attachment aa = question.getAttachment();
+            question.setAttachment(null);
+            question.save();
+            aa.delete();
+            // DO NOT DELETE THE ACTUAL FILE, IT MAY BE REFERENCED FROM CHILD QUESTIONS!
+            //AppUtil.removeAttachmentFile(aa.getFilePath());
+        }
         return redirect("/#/questions/" + String.valueOf(id));
     }
 
     @Restrict({@Group("TEACHER"), @Group("ADMIN"), @Group("STUDENT")})
     public Result deleteQuestionAnswerAttachment(Long qid, String hash) {
-
         Question question = Ebean.find(Question.class, qid);
-        Attachment aa = Ebean.find(Attachment.class, question.getAnswer().getAttachment().getId());
-
-        Answer answer = question.getAnswer();
-        answer.setAttachment(null);
-        answer.save();
-        aa.delete();
-        AppUtil.removeAttachmentFile(aa.getFilePath());
-
+        if (question != null && question.getAnswer() != null && question.getAnswer().getAttachment() != null) {
+            Answer answer = question.getAnswer();
+            Attachment aa = answer.getAttachment();
+            answer.setAttachment(null);
+            answer.save();
+            AppUtil.removeAttachmentFile(aa.getFilePath());
+            aa.delete();
+        }
         return redirect("/#/student/doexam/" + hash);
     }
 
@@ -149,16 +164,31 @@ public class AttachmentController extends BaseController {
     public Result deleteExamAttachment(Long id) {
         Exam exam = Ebean.find(Exam.class, id);
         if (exam.getAttachment() != null) {
-            Attachment aa = Ebean.find(Attachment.class, exam.getAttachment().getId());
-            if (aa != null) {
-                exam.setAttachment(null);
-                exam.save();
-                aa.delete();
-                // DO NOT DELETE THE ACTUAL FILE, IT MAY BE REFERENCED FROM CHILD EXAMS!
-                // AppUtil.removeAttachmentFile(aa.getFilePath());
-            }
+            Attachment aa = exam.getAttachment();
+            exam.setAttachment(null);
+            exam.save();
+            aa.delete();
+            // DO NOT DELETE THE ACTUAL FILE, IT MAY BE REFERENCED FROM CHILD EXAMS!
+            // AppUtil.removeAttachmentFile(aa.getFilePath());
         }
         return redirect("/#/exams/" + String.valueOf(id));
+    }
+
+    @Restrict({@Group("TEACHER"), @Group("ADMIN")})
+    public Result deleteFeedbackAttachment(Long id) {
+        Exam exam = Ebean.find(Exam.class, id);
+        if (exam == null) {
+            return notFound("sitnet_exam_not_found");
+        }
+        Comment comment = exam.getExamFeedback();
+        if (comment != null && comment.getAttachment() != null) {
+            Attachment aa = comment.getAttachment();
+            comment.setAttachment(null);
+            comment.save();
+            AppUtil.removeAttachmentFile(aa.getFilePath());
+            aa.delete();
+        }
+        return ok();
     }
 
     @Restrict({@Group("TEACHER"), @Group("ADMIN")})
@@ -184,6 +214,13 @@ public class AttachmentController extends BaseController {
         } catch (IOException e) {
             return internalServerError("sitnet_error_creating_attachment");
         }
+        // Delete existing if exists
+        if (exam.getAttachment() != null) {
+            Attachment aa = exam.getAttachment();
+            exam.setAttachment(null);
+            exam.save();
+            aa.delete();
+        }
 
         Attachment attachment = new Attachment();
         attachment.setFileName(filePart.getFilename());
@@ -197,16 +234,64 @@ public class AttachmentController extends BaseController {
         return ok(attachment);
     }
 
+    @Restrict({@Group("TEACHER"), @Group("ADMIN")})
+    public Result addFeedbackAttachment(Long id) {
+        MultipartFormData body = request().body().asMultipartFormData();
+        FilePart filePart = body.getFile("file");
+        if (filePart == null) {
+            return notFound();
+        }
+        File file = filePart.getFile();
+        if (file.length() > AppUtil.getMaxFileSize()) {
+            return forbidden("sitnet_file_too_large");
+        }
+        Exam exam = Ebean.find(Exam.class, id);
+        if (exam == null) {
+            return notFound();
+        }
+        if (exam.getExamFeedback() == null) {
+            Comment comment = new Comment();
+            AppUtil.setCreator(comment, getLoggedUser());
+            comment.save();
+            exam.setExamFeedback(comment);
+            exam.update();
+        }
+        String newFilePath;
+        try {
+            newFilePath = copyFile(file, "exam", id.toString(), "feedback");
+        } catch (IOException e) {
+            return internalServerError("sitnet_error_creating_attachment");
+        }
+        Comment comment = exam.getExamFeedback();
+        // Delete old one if exists
+        if (comment.getAttachment() != null) {
+            Attachment aa = comment.getAttachment();
+            comment.setAttachment(null);
+            comment.save();
+            AppUtil.removeAttachmentFile(aa.getFilePath());
+            aa.delete();
+        }
+        Attachment attachment = new Attachment();
+        attachment.setFileName(filePart.getFilename());
+        attachment.setFilePath(newFilePath);
+        attachment.setMimeType(filePart.getContentType());
+        attachment.save();
+
+        comment.setAttachment(attachment);
+        comment.save();
+
+        return ok(attachment);
+    }
+
     @Restrict({@Group("TEACHER"), @Group("ADMIN"), @Group("STUDENT")})
     public Result downloadQuestionAttachment(Long id) {
 
-        Question question = Ebean.find(Question.class)
-                .fetch("attachment")
-                .where()
-                .eq("id", id)
-                .findUnique();
+        Question question = Ebean.find(Question.class, id);
+        if (question == null || question.getAttachment() == null) {
+            return notFound();
+        }
 
-        Attachment aa = Ebean.find(Attachment.class, question.getAttachment().getId());
+        Attachment aa = question.getAttachment();
         File file = new File(aa.getFilePath());
 
         response().setHeader("Content-Disposition", "attachment; filename=\"" + aa.getFileName() + "\"");
@@ -216,13 +301,11 @@ public class AttachmentController extends BaseController {
     @Restrict({@Group("TEACHER"), @Group("ADMIN"), @Group("STUDENT")})
     public Result downloadQuestionAnswerAttachment(Long qid, String hash) {
 
-        Question question = Ebean.find(Question.class)
-                .fetch("attachment")
-                .where()
-                .eq("id", qid)
-                .findUnique();
-
-        Attachment aa = Ebean.find(Attachment.class, question.getAnswer().getAttachment().getId());
+        Question question = Ebean.find(Question.class, qid);
+        if (question == null || question.getAnswer() == null || question.getAnswer().getAttachment() == null) {
+            return notFound();
+        }
+        Attachment aa = question.getAnswer().getAttachment();
         File file = new File(aa.getFilePath());
 
         response().setHeader("Content-Disposition", "attachment; filename=\"" + aa.getFileName() + "\"");
@@ -232,13 +315,25 @@ public class AttachmentController extends BaseController {
     @Restrict({@Group("TEACHER"), @Group("ADMIN"), @Group("STUDENT")})
     public Result downloadExamAttachment(Long id) {
 
-        Exam exam = Ebean.find(Exam.class)
-                .fetch("attachment")
-                .where()
-                .eq("id", id)
-                .findUnique();
+        Exam exam = Ebean.find(Exam.class, id);
+        if (exam == null || exam.getAttachment() == null) {
+            return notFound();
+        }
+        Attachment aa = exam.getAttachment();
+        File file = new File(aa.getFilePath());
 
-        Attachment aa = Ebean.find(Attachment.class, exam.getAttachment().getId());
+        response().setHeader("Content-Disposition", "attachment; filename=\"" + aa.getFileName() + "\"");
+        return ok(com.ning.http.util.Base64.encode(setData(file).toByteArray()));
+    }
+
+    @Restrict({@Group("TEACHER"), @Group("ADMIN"), @Group("STUDENT")})
+    public Result downloadFeedbackAttachment(Long id) {
+
+        Exam exam = Ebean.find(Exam.class, id);
+        if (exam == null || exam.getExamFeedback() == null || exam.getExamFeedback().getAttachment() == null) {
+            return notFound();
+        }
+        Attachment aa = exam.getExamFeedback().getAttachment();
         File file = new File(aa.getFilePath());
 
         response().setHeader("Content-Disposition", "attachment; filename=\"" + aa.getFileName() + "\"");
